requirements.txt
sqlite3
pyyaml
hades
